corta deps e reorganiza back
