+++ back/src/bd/queries.ts
